requirements.txt
streamlit>=1.35.0
tensorflow>=2.10.0
pillow
numpy
requests
